implement logic in inventoryStatus endpoint

+++ server/src/main/java/dev/kjj/pharmaphix/dtos/InventoryResponseDto.java
@@ -1,4 +1,4 @@
 package dev.kjj.pharmaphix.dtos;
 
-public record InventoryResponseDto(int needToBeOrdered) {
+public record InventoryResponseDto(long needToBeOrdered) {
 }
